split bytes.ts into bytes.ts and buffers.ts
